docs(root): refactor omni_example to index and debugging.ts
BTC-1864

TICKET: BTC-0
